+++ CHANGES.md
@@ -1,6 +1,7 @@
 ### Apache MetaModel 5.2.1 (WIP)
 
  * [METAMODEL-1208] - Fixed closing of JDBC ResultSets when getting certain metadata.
+ * [METAMODEL-156] - Fixed issue in selection of default schema when multiple schema names match substring of connection string.
 
 ### Apache MetaModel 5.2.0